Unified `parallel --bar` syntax using Bash's process substitution.
In both cases - when WordNet IDs are specified on the command-line, and
in a file - Parallel is invoked using a command of the form
```bash
	parallel --bar 'download_wnid {}' < <wordnet-ids>
```
where `<wordnet-ids>` contains the input WordNet IDs.
Earlier, in the former case, we were using a command of the form
```bash
	<wordnet-ids> | parallel --bar 'download_wnid {}'
```
i.e. the WordNet IDs were getting piped into Parallel.

The end result is the same, this just makes the code cleaner.